spike: moving circ dep check to pre-commit
